Split FileUpload queue renderer
