+++ packages/components/src/components/file-upload/file-upload.svelte
@@ -21,16 +21,14 @@
 </script>
 
 <script lang="ts">
-  import RotateCcwIcon from 'lucide-svelte/icons/rotate-ccw';
-  import XIcon from 'lucide-svelte/icons/x';
-
   import { resolveFieldControl } from '../../_internal/field-control.ts';
   import { getFormFieldContext } from '../../_internal/form-field-context.ts';
   import { classNames } from '../../utilities/class-names.ts';
   import { devWarn } from '../../utilities/dev-warn.ts';
   import { formatBytes } from '../../utilities/format-bytes.ts';
   import { useAnnouncer } from '../../utilities/use-announcer.svelte.ts';
-  import { acceptsFile, fileTypeIcon, formatAcceptDescription } from './file-upload-utilities.ts';
+  import FileUploadList from './file-upload-list.svelte';
+  import { acceptsFile, formatAcceptDescription } from './file-upload-utilities.ts';
   import type { FileUploadEntry, FileUploadProps, RejectedFile } from './file-upload.types.ts';
 
   const INTERACTIVE_ARIA_ROLES = [
@@ -357,11 +355,6 @@
     onFilesChange?.(nextEntries);
     announcer.announce(`${entry.file.name} removed`);
   }
-
-  function progressValue(progress: number | undefined): number {
-    if (progress === undefined) return 0;
-    return Math.max(0, Math.min(100, progress));
-  }
 </script>
 
 {#snippet defaultIdle()}
@@ -458,117 +451,13 @@
     {#if fileList}
       {@render fileList(renderedEntries)}
     {:else}
-      <ul class="cinder-file-upload__list">
-        {#each renderedEntries as entry (entry.id)}
-          {@const errorId = entry.error ? `${resolvedId}-${entry.id}-error` : undefined}
-          {@const FileTypeIcon = fileTypeIcon(entry.file.type)}
-          <li class="cinder-file-upload__row" aria-describedby={errorId}>
-            <div class="cinder-file-upload__row-main">
-              <div class="cinder-file-upload__file-details">
-                <FileTypeIcon class="cinder-file-upload__file-icon" aria-hidden="true" />
-                <div class="cinder-file-upload__file-meta">
-                  <span class="cinder-file-upload__file-name cinder-_truncate"
-                    >{entry.file.name}</span
-                  >
-                  <span class="cinder-file-upload__file-size">{formatBytes(entry.file.size)}</span>
-                </div>
-              </div>
-
-              <div class="cinder-file-upload__row-actions">
-                {#if entry.status === 'uploading'}
-                  <span class="cinder-file-upload__status" data-status="uploading">Uploading</span>
-                {:else if entry.status === 'success'}
-                  <span class="cinder-file-upload__status" data-status="success">
-                    <svg
-                      class="cinder-file-upload__status-icon"
-                      aria-hidden="true"
-                      viewBox="0 0 16 16"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M3.5 8.5L6.25 11.25L12.5 5"
-                        stroke="currentColor"
-                        stroke-width="1.5"
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                      />
-                    </svg>
-                    Complete
-                  </span>
-                {:else if entry.status === 'error'}
-                  <span class="cinder-file-upload__status" data-status="error">
-                    <svg
-                      class="cinder-file-upload__status-icon"
-                      aria-hidden="true"
-                      viewBox="0 0 16 16"
-                      fill="none"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        d="M8 4.5V8.25M8 11H8.00667M14 8C14 11.3137 11.3137 14 8 14C4.68629 14 2 11.3137 2 8C2 4.68629 4.68629 2 8 2C11.3137 2 14 4.68629 14 8Z"
-                        stroke="currentColor"
-                        stroke-width="1.5"
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                      />
-                    </svg>
-                    Failed
-                  </span>
-                {:else}
-                  <span class="cinder-file-upload__status" data-status="pending">Pending</span>
-                {/if}
-                <button
-                  type="button"
-                  class="cinder-file-upload__remove"
-                  disabled={field.disabled}
-                  aria-label={`Remove ${entry.file.name}`}
-                  onclick={() => removeEntry(entry)}
-                >
-                  <XIcon aria-hidden="true" />
-                </button>
-              </div>
-            </div>
-
-            {#if entry.status === 'uploading'}
-              {@const value = progressValue(entry.progress)}
-              <div
-                class="cinder-file-upload__progress"
-                role="progressbar"
-                aria-label={`Uploading ${entry.file.name}`}
-                aria-valuemin="0"
-                aria-valuemax="100"
-                aria-valuenow={value}
-              >
-                <div
-                  class="cinder-file-upload__progress-fill"
-                  style={`--cinder-file-upload-progress: ${value}`}
-                ></div>
-              </div>
-            {/if}
-
-            {#if entry.error || (entry.status === 'error' && entry.rejectionReason === undefined && onRetry)}
-              <div class="cinder-file-upload__error-row">
-                {#if entry.error}
-                  <p id={errorId} class="cinder-file-upload__error">{entry.error}</p>
-                {/if}
-                {#if entry.status === 'error' && entry.rejectionReason === undefined && onRetry}
-                  <button
-                    type="button"
-                    class="cinder-file-upload__retry"
-                    aria-label={`Retry ${entry.file.name}`}
-                    aria-describedby={errorId}
-                    onclick={() => onRetry(entry)}
-                  >
-                    <RotateCcwIcon class="cinder-file-upload__retry-icon" aria-hidden="true" />
-                    Retry
-                  </button>
-                {/if}
-              </div>
-            {/if}
-          </li>
-        {/each}
-      </ul>
+      <FileUploadList
+        entries={renderedEntries}
+        {resolvedId}
+        disabled={field.disabled}
+        {onRetry}
+        onRemove={removeEntry}
+      />
     {/if}
   {/if}
